+++ resources/react/components/UploadPreview.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const UploadPreview = () => {
+    return (
+        <>
+
+        </>
+    )
+}
+
+export default UploadPreview