+++ stubs/app/Models/User.php
@@ -2,20 +2,22 @@
 
 namespace App\Models;
 
-use DesignByCode\Guardian\Http\Traits\AvatarTrait;
+use DesignByCode\Guardian\Http\Traits\Avatar;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Fortify\TwoFactorAuthenticatable;
+use Spatie\MediaLibrary\HasMedia;
 
-class User extends Authenticatable implements MustVerifyEmail
+
+class User extends Authenticatable implements MustVerifyEmail, HasMedia
 {
     use HasFactory;
     use Notifiable;
     use TwoFactorAuthenticatable;
-    use AvatarTrait;
-
+    use Avatar;
+    
     /**
      * The attributes that are mass assignable.
      *